Made an array-declaration in Movie.h clearer to where the number comes from.
Commented out some movie backwards compatibility cases dependent on the old movie version system.

Improved unsavedMovieData detection when opening a new movie when a movie is already opened. This should work better but it may still be unreliable. The downside is that the edit box for movie file name had to be made Read Only, now Browse has to be used to pick a new movie, even for smaller typos, which is closer to how emulators does it, so there should not be any added usability complexity.