WIP - Begin DbgHelp wrapper to obsolete extendedtrace
This is a work in progress, the goal is to get rid of the DbgHelp library and the extendedtrace code. This is 3rd party code that's not really well adapted for our needs. Let's replace this with code that is more adapted to C++ idioms, and that is easily extendable as our needs grow.

The purpose, besides keeping the extendedtrace functionality, is to:

*  Make it "easy" to implement a code viewer and tracer of the DLL in Hourglass so that we can debug ourselves, currently we can only really printf-debug the DLL, which is horrible.
*  Allow to throw out the trustedcaller-hack in the DLL and make it an debugger-command which can do proper evaluation based on a stack trace.

The latter will cost speed, but the hack is really shitty, and it has to go, there's no way around that, as the hack requires coder-prediction of the call chain, which makes it very hard to properly implement some of the checks in a clean and easy-to-understand manner.

Note: DIA requires unicode to prevent all interactions with the library to be super-painful. I migrated to unicode haphazardly throughout the process of adding DIA, so it's really baked into everywhere. Sorry =/

Known issues:

*  PE header parsing is broken for some DLLs, also, the code should move.
*  The interface to the debug core is incomplete, it misses everything that is useful.